#22 [FEAT] signup_basicinfo_view_end
