+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.nadosunbae_andorid">
+    package="com.nadosunbae_andorid" >
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.NadoSunbaeAndroid">
+        android:theme="@style/Theme.NadoSunbaeAndroid" >
+        <activity
+            android:name=".presentation.ui.sign.SignInCustomDialog"
+            android:exported="true" />
         <activity
             android:name=".presentation.ui.sign.SignUpFinishActivity"
             android:exported="true" />
@@ -26,11 +29,10 @@
             android:exported="true" />
         <activity
             android:name=".presentation.ui.sign.FindPwActivity"
-            android:exported="true"/>
+            android:exported="true" />
         <activity
             android:name=".presentation.ui.sign.SignInActivity"
-            android:exported="true"/>
-
+            android:exported="true" />
         <activity
             android:name=".presentation.ui.mypage.ServiceInqueryActivity"
             android:exported="false" />
@@ -55,13 +57,12 @@
         <activity
             android:name=".presentation.ui.mypage.MyPagePostActivity"
             android:exported="false" />
-
         <activity
             android:name=".presentation.ui.review.ReviewWriteActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.ui.main.MainActivity"
-            android:exported="true">
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
