+++ app/src/main/java/com/nadosunbae_andorid/util/CustomDialog.kt
@@ -7,42 +7,8 @@ import android.view.View
 import android.view.WindowManager
 import com.nadosunbae_andorid.R
 import com.nadosunbae_andorid.presentation.ui.sign.SignUpBasicInfoActivity
-import kotlinx.android.synthetic.main.signup_custom_dialog.*
 
 class CustomDialog(context : Context) {
-
-    private val dialog = Dialog(context)
-    private lateinit var onClickedListener : ButtonClickListener
-
-
-
-
-    interface ButtonClickListener{
-        fun onClicked(num: () -> Unit)
-    }
-
-    fun setOnClickListener(listener: ButtonClickListener)
-    {
-        onClickedListener = listener
-    }
-
-    fun showDialog(){
-
-        dialog.setContentView(R.layout.signup_custom_dialog)
-        dialog.window?.setLayout(WindowManager.LayoutParams.WRAP_CONTENT, WindowManager.LayoutParams.WRAP_CONTENT)
-        dialog.show()
-
-        dialog.text_signup_dialog_out.setOnClickListener {
-            onClickedListener.onClicked{
-                SignUpBasicInfoActivity().closePage()
-            }
-            dialog.dismiss()
-        }
-
-        dialog.text_signup_dialog_in.setOnClickListener {
-            dialog.dismiss()
-        }
-    }
