+++ app/src/main/res/layout/activity_sign_in_custom_dialog.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+  <data>
+
+  </data>
+
+  <androidx.constraintlayout.widget.ConstraintLayout
+      android:id="@+id/cl_signup_costom_dialog"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:background="@drawable/rectangle_fill_white_8dp">
+
+
+    <TextView
+        android:id="@+id/text_signup_dialog"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="34dp"
+        android:fontFamily="@font/pretendard_regular"
+        android:text="@string/signup_alert"
+        android:textAlignment="center"
+        android:textColor="@color/main_text"
+        android:textSize="14sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/text_signup_dialog_out"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="14dp"
+        android:layout_marginTop="24dp"
+        android:layout_marginBottom="14dp"
+        android:background="@drawable/rectangle_fill_gray0_12dp"
+        android:fontFamily="@font/pretendard_regular"
+        android:paddingLeft="32dp"
+        android:paddingTop="15dp"
+        android:paddingRight="32dp"
+        android:paddingBottom="15dp"
+        android:text="@string/signup_alert_out"
+        android:textColor="@color/gray_3"
+        android:textSize="14sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/text_signup_dialog" />
+
+
+    <TextView
+        android:id="@+id/text_signup_dialog_in"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="14dp"
+        android:layout_marginEnd="14dp"
+        android:background="@drawable/rectangle_fill_black_12dp"
+        android:fontFamily="@font/pretendard_regular"
+        android:paddingLeft="32dp"
+        android:paddingTop="15dp"
+        android:paddingRight="32dp"
+        android:paddingBottom="15dp"
+        android:text="@string/signup_alert_in"
+        android:textColor="@color/main_default"
+        android:textSize="14sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/text_signup_dialog_out"
+        app:layout_constraintTop_toTopOf="@+id/text_signup_dialog_out" />
+
+  </androidx.constraintlayout.widget.ConstraintLayout>
+
+</layout>
